Move the Forger class into its own module
Move the `Forger` class into its own module.

* **New Module: `forger.py`**
  - Create a new module `forger.py` and move the `Forger` class from `tinychain.py` to this new module.
  - Import necessary dependencies in the new module.

* **Changes in `tinychain.py`**
  - Remove the `Forger` class from `tinychain.py`.
  - Import the `Forger` class from the new `forger` module.
  - Update the instantiation of the `Forger` class to use the new import.
